+++ fs/axfs-ng-vfs/src/node/file.rs
@@ -98,9 +98,6 @@ pub trait FileNodeOps: NodeOps + FsPollable {
         Err(VfsError::OperationNotSupported)
     }
 
-    /// Sets the file's symlink target.
-    fn set_symlink(&self, target: &str) -> VfsResult<()>;
-
     /// Manipulates the underlying device parameters of special files.
     fn ioctl(&self, _cmd: u32, _arg: usize) -> VfsResult<usize> {
         Err(VfsError::NotATty)